extract ready screen into a separate class
